♻️ refactor: lendingcard
