+++ views/defi/components/all/all.data.ts
@@ -1,0 +1,20 @@
+export const lendingOptions = [
+  {
+    platform: 'Nemo',
+    badgeColor: '#3366FF',
+    asset: 'WAL-SUI',
+    apr: '2.21%',
+    tvl: '$66.4K',
+    category: 'Lending',
+    points: '0.033/WAL/day',
+  },
+  {
+    platform: 'Bucket',
+    badgeColor: '#1A1A1A',
+    asset: 'WAL-SUI',
+    apr: '2.21%',
+    tvl: '$66.4K',
+    category: 'Lending',
+    points: '--',
+  },
+];
