+++ views/defi/components/all/components/info-box/info-box.types.ts
@@ -1,0 +1,7 @@
+import { ReactNode } from 'react';
+
+export interface IInfoBox {
+  label: string;
+  value: string;
+  icon?: ReactNode;
+}
