+++ views/defi/components/all/components/lending-badge/lending-badge.types.ts
@@ -1,0 +1,4 @@
+export interface ILendingBadge {
+  platform: string;
+  badgeColor: string;
+}
