+++ views/defi/components/all/components/lending-card/lending-card.types.ts
@@ -1,0 +1,9 @@
+export interface ILendingCard {
+  platform: string;
+  badgeColor: string;
+  asset: string;
+  apr: string;
+  tvl: string;
+  category: string;
+  points: string;
+}
